requirements.txt
flask
gunicorn
requests
fuzzywuzzy
python-Levenshtein
inflect
python-dotenv
symspellpy
flask-cors